Migrate person header and remove sort code
