+++ shared/package.json
@@ -20,6 +20,8 @@
   "dependencies": {
     "@emotion/react": "11.10.5",
     "@emotion/styled": "11.10.5",
+    "@mantine/core": "5.9.5",
+    "@mantine/hooks": "5.9.5",
     "@mui/material": "5.11.1",
     "@mui/system": "5.11.1",
     "md5": "2.3.0",
