--- shared/src/components/Persons/constants/sort.ts
@@ -1,9 +1,0 @@
-export enum SORT_TYPE {
-  alphabetically = 'Alphabetically',
-  bookmarks = 'Bookmarks',
-}
-
-export enum SORT_ORDER {
-  asc = 'Ascending',
-  desc = 'Descending',
-}
