+++ shared/src/components/Persons/utils/index.ts
@@ -1,6 +1,5 @@
 import { IPerson, IPersons } from '../interfaces/persons';
 import { hasText } from '../../../utils/search';
-import { SORT_ORDER } from '../constants/sort';
 
 export const decodePerson = (person: IPerson): IPerson => {
   const { uid, imageRef, name, taggedUrls } = person;
@@ -23,13 +22,7 @@ export const getReactKey = (row: number, column: number, gridSize: number) =>
 export const getFilteredPersons = (persons: IPerson[], searchText: string) =>
   persons.filter(({ name }) => !searchText || hasText(searchText, name));
 
-export const sortAlphabetically = <T extends IPerson>(
-  sortOrder: SORT_ORDER,
-  persons: T[]
-) => {
+export const sortAlphabetically = <T extends IPerson>(persons: T[]) => {
   const sortedPersons = persons.sort((a, b) => a.name.localeCompare(b.name));
-  if (sortOrder === SORT_ORDER.desc) {
-    sortedPersons.reverse();
-  }
   return [...sortedPersons];
 };
